Remove test function and optimize compute_time_segments types

+++ src/mims_helper.c
@@ -43,9 +43,9 @@ static void compute_time_segments(
     dataframe->n_segments = (uint16_t)((dataframe->timestamps[dataframe->size - 1] - start_time) / (double)break_size_in_seconds) + 1;
     dataframe->segments = malloc(dataframe->n_segments * sizeof(uint16_t));
 
-    int segment_i = 0;
-    int last_break_index = 0;
-    for (int i = 0; i < dataframe->size; i++)
+    uint16_t segment_i = 0;
+    uint32_t last_break_index = 0;
+    for (uint32_t i = 0; i < dataframe->size; i++)
     {
         if (((dataframe->timestamps[i] - dataframe->timestamps[last_break_index]) >= break_size_in_seconds) || (i == dataframe->size - 1))
         {
